requirements.txt
# API Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
torch==2.1.2
pytesseract==0.3.10
PyPDF2==3.0.1
Pillow==10.2.0
opencv-python==4.9.0.80

# Scheduling
APScheduler==3.10.4

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
joblib==1.3.2
tqdm==4.66.1
loguru==0.7.2

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
